requirements.txt
# Datapizza-ai framework
#
datapizza-ai
datapizza-ai-clients-openai
datapizza-ai-clients-openai-like

# Others
python-dotenv
PyYAML

# UI
streamlit